token

+++ Back-end/src/controllers/products/index.js
@@ -1,4 +1,7 @@
 const Products = require('../../models/product/product')
+const User = require('../../models/user/user')
+const { getTokenData} = require('../../jwtAuth.js')
+const { SECRET_TOKEN } = process.env
 
 module.exports ={
 
@@ -66,52 +69,93 @@ module.exports ={
      },
       
     postProducts:async (req,res)=>{
-        const {name, description, price, image, category, count, color} = req.body
-      if(!name){
-        return res.status(404).send('Falta información, el NOMBRE esta incompleta')
-      }
-      else if(!description){
-        return res.status(404).send('Falta información, la DESCRIPCION esta incompleta')
-      }
-      else if(!price){
-        return res.status(404).send('Falta información, el PRECIO esta incompleta')
-      }
-      else if(!image){
-        return res.status(404).send('Falta información, la IMAGEN esta incompleta')
-      }
-      else if(!category){
-        return res.status(404).send('Falta información, la CATEGORIA esta incompleta')
-      }
-      else if(!count){
-        return res.status(404).send('Falta información, la CANTIDAD esta incompleta')
-      }
-      else if(!color){
-        return res.status(404).send('Falta información, los COLORES esta incompleta')
-      }
-      else{
-        const product = new Products({
-            name,
-            description,      
-            price,
-            image,
-            category,
-            count,
-            color
+    const { name, description, price, image, category, count, color } = req.body
+
+    const autorization = req.get('authorization')
+    if (!autorization) {
+      return res.status(401).json({ message: 'No tienes permisos para hacer esto' })
+    }
+    if (autorization.split(' ')[0].toLowerCase() !== 'bearer') {// bearer zxaUHUHU. este if esta verificando que exista el bearer
+      return res.status(401).json({ message: 'No tienes permisos para hacer esto' })
+    }
+    const token = autorization.substring(7)//esta costante va a contener el token, el token principalmente es asi bearer jdsiijyVGVG, Y CON EL SUBSTRING(7) saca a bearer y deja el token solo
+    const data = getTokenData(token) // le mandatmos a la funcion getTokenData el token que nos pasaron eso nos va a responder con la data o un error 
+    //console.log(data)
+    if (!data) { 
+      return res.status(401).json({ message: 'No tienes permisos para hacer esto' })
+    }
+    const user = await User.findById(data.id) //si nos manda un token no valido  hacemos una verificacion si existe algun usuario con el id 
+    if(!user) {
+      return res.status(404).json({ message: 'No se ha encontrado usuario' })
+    }
+
+
+
+    if (!name) {
+      return res.status(404).send('Falta información, el NOMBRE esta incompleta')
+    }
+    else if (!description) {
+      return res.status(404).send('Falta información, la DESCRIPCION esta incompleta')
+    }
+    else if (!price) {
+      return res.status(404).send('Falta información, el PRECIO esta incompleta')
+    }
+    else if (!image) {
+      return res.status(404).send('Falta información, la IMAGEN esta incompleta')
+    }
+    else if (!category) {
+      return res.status(404).send('Falta información, la CATEGORIA esta incompleta')
+    }
+    else if (!count) {
+      return res.status(404).send('Falta información, la CANTIDAD esta incompleta')
+    }
+    else if (!color) {
+      return res.status(404).send('Falta información, los COLORES esta incompleta')
+    }
+    else {
+      const product = new Products({
+        name,
+        description,
+        price,
+        image,
+        category,
+        count,
+        color
+      })
+      product.save()
+        .then(() => {
+          return res.status(201).json({ msg: 'Producto creado', product })
         })
-        product.save()
-        .then(()=>{
-          return res.status(201).json({msg:'Producto creado', product})
-            })
-        .catch((error)=>{
+        .catch((error) => {
           return res.status(500).send(error)
         })
     }
-},
+  },
 
   updateProduct: async (req,res)=>{
     try {
       const {name, description, price, image, category, count, color} = req.body
       const {id} = req.params
+    
+      const authorization = req.get('authorization')
+      if(!authorization){
+        return res.status(401).json({ message: 'No tienes permisos para hacer esto' })
+      }
+      if(authorization.split(' ')[1].toLowerCase() !== 'bearer'){
+        return res.status(401).json({ message: 'No tienes permisos para hacer esto' })
+      }
+      const token = authorization.substring(7)
+      let data = getTokenData(token)
+      if(!data){
+        return res.status(401).json({ message: 'No tienes permisos para hacer esto' })
+      }
+      const userToken = await User.findById(data.id)
+      if(!userToken){
+        return res.status(404).json({ message: 'No se ha encontrado usuario' })
+      }
+
+
+
       if(!name){
         return res.status(404).send('Falta información, el NOMBRE esta incompleta')
       }
@@ -145,6 +189,25 @@ module.exports ={
   deleteProducts: async (req, res)=>{
     try {
       const {id}= req.params
+      
+      const authorization = req.get('authorization')
+      if(!authorization){
+        return res.status(401).json({ message: 'No tienes permisos para hacer esto' })
+      }
+      if(authorization.split(' ')[0].toLowerCase() !== "bearer"){
+        return res.status(401).json({ message: 'No tienes permisos para hacer esto' })
+      }
+      let token = authorization.split(' ')[1]
+      let data = getTokenData(token)
+      if(!data){
+        return res.status(401).json({ message: 'No tienes permisos para hacer esto' })
+      }
+      const userToken = await User.findById(data.id)
+      if(!userToken){
+        return res.status(404).json({ message: 'No se ha encontrado usuario' })
+      }
+
+      
       if(id){
        const deletee = await Products.findByIdAndDelete(id);
        res.json({msg:'Producto eliminado', deletee})
